new version of graying

+++ gray-scale/src/gray/desaturation.rs
@@ -12,7 +12,7 @@ impl Gray for Desaturation{
     fn gray(&self, red: u8, blue: u8, green: u8)-> (u8, u8, u8){
         let max = cmp::max(cmp::max(red, green), blue);
         let min = cmp::min(cmp::min(red, green), blue);
-        let color =  (min as u16 + max as u16)  as f32 /2.0;
+        let color =  (min as u32 + max as u32)  as f32 /2.0;
         (color as u8, color as u8, color as u8)
     }
 }
